tests/requirements.txt
boto3==1.9.250
awscli==1.16.260
pytest==5.2.1
pylint==2.3.1
python-box==3.2.4
